+++ rio-window/src/platform/macos.rs
@@ -40,6 +40,9 @@ pub trait WindowExtMacOS {
     /// Sets whether or not the window has shadow.
     fn set_has_shadow(&self, has_shadow: bool);
 
+    /// Sets background color.
+    fn set_background_color(&self, _r: f64, _g: f64, _b: f64, _a: f64);
+
     /// Group windows together by using the same tabbing identifier.
     ///
     /// <https://developer.apple.com/documentation/appkit/nswindow/1644704-tabbingidentifier>
@@ -114,6 +117,12 @@ impl WindowExtMacOS for Window {
     fn set_has_shadow(&self, has_shadow: bool) {
         self.window
             .maybe_queue_on_main(move |w| w.set_has_shadow(has_shadow))
+    }
+
+    #[inline]
+    fn set_background_color(&self, r: f64, g: f64, b: f64, a: f64) {
+        self.window
+            .maybe_queue_on_main(move |w| w.set_background_color(r, g, b, a))
     }
 
     #[inline]